認証システムを削除する
