+++ package.json
@@ -10,7 +10,6 @@
   },
   "dependencies": {
     "next": "15.2.1",
-    "next-auth": "^4.24.11",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
