+++ src/app/layout.tsx
@@ -4,8 +4,6 @@
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 
-import { SessionProvider } from "next-auth/react"
-
 const geistSans = Geist({
   variable: "--font-geist-sans",
   subsets: ["latin"],
@@ -22,14 +20,12 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <SessionProvider>
-      <html lang="en">
+    <html lang="en">
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
         {children}
       </body>
     </html>
-    </SessionProvider>
   );
 }
